app/requirements.txt
pyautogui
pillow
tkcap
pydicom
img2pdf
#opencv_python
#matplotlib
#pandas
tensorflow
#python-xlib

numpy
opencv-python
python-xlib; sys_platform == "linux"

